Fix I.Composition proofs broken by linkability
The proofs are still quite brittle, but can be fixed with relative ease
by noting the sequence of inversions that introduce the new
well-formedness hypotheses and offsetting two by two blockwise.

Linkability with the empty program is no longer a theorem, so a bit of
extra work is needed to inline the relevant parts. Other case analyses
in the partially admitted proofs are no longer applicable.