Revised RSC_DC_MD theorem and prefix lemmas
Pushing the new traces to the statements of the various lemmas in
previous commits, the changes to the main proof are quite modest.

Lemmas on prefixes and related asserts must be rephrased in term of the
new relationships between types of prefixes, behaviors and traces. In
particular, this asymmetry extends to [behavior_prefix_comp]. Other
lemma statements are modified to match the context as it exists in the
proof, leaving those proofs to be amended. The asymmetry also causes
transitivity lemmas on prefixes to split, as they are no longer applied
to two traces-as-prefixes, but to a mixture of types.