Fix treatment of prefixes (I)
By inspecting the new kinds of prefix information available to us, we
begin to replace instances of problematic lemmas with sound reasoning.
In this instance, a lemma whose revision was tainted is replaced with a
simpler, sound lemma, which can be applied in the main proof after some
additional preprocessing.