Fix treatment of prefixes (II)
Again by analyzing the information in the new prefixes, the sanitization
of the parts of the main proof that relied on tainted lemmas is
completed. The statement of the soured lemma, and the lemma that
depended on it, are removed.

Although the proof of [behavior_prefix_comp] needs to be redone, this is
a matter of adapting Carmine's scheme to the new prefixes and is not
problematic.