Extract traces from finite behavior prefixes
In the revised proof if definability with linking, behavior prefixes
(a la CompCert) change and are now indirectly wrapped in finite behavior
prefixes. Most of the proof goes through by extraction of the trace from
the finite behavior prefix (one [have] block remains to be treated in
the next commit).