Weak separate compilation in RSC_DC_MD proof
The FCC assertion is modified to state not a strict correspondence
between compilation as the sum of the compiled parts, but the existence
of a behaviorally equivalent compilation, that is to say the novelty in
the weaker statement of separate compilation, which is used to prove it.
Further assertions about the behavior of the result of compilation are
rewritten to use the new, behaviorally equivalent compiled program
instead of the composition of the compiled parts.

The BCC assertion operates in the same fashion, although in its current
structure it needs these behavioral requirements to be stated ahead of
the main assert.

The only details that remain open after these changes are the conditions
under which the compiled programs can be generated, which should be
straightforward from the context.