Propagate mains to high-level RSC_DC_MC proof
See comments for missing pieces, which probably need to come from
definability.

Ancillary changes would concern the definitions of linkable_mains. The
current injection that occurs in several places, based on the definition
in the Linking module, becomes messy to work with because it is quite
low-level.

The underlying issue may be the absence of main-related information in
the interfaces, which a priori forces us to impose additional hypotheses
threaded throughout the proof.