Strong state determinism on program (II)
Add a second lemma on the preservation of the top of the partial stack.
Using this and the previous lemma, obtain two higher-level results about
the inequality of the top of the partial stacks obtained by
contradiction. Use these two lemmas in turn to discharge all nonsensical
sub-goals in the proof of strong state determinism on programs.