High-level proof of star_improvement
The extant high-level proof was based on starN and a case analysis on
the relation between the number of steps in the two stars, with an
auxiliary lemma given, but not proved, for the case where the number of
steps is identical.

While there is nothing wrong with the statement of said auxiliary lemma,
which mimics closely the structure of the proof of star_improvement (for
the case of interest), it is questionable that the use of starN is of
assistance in the proof. In particular, the presence of epsilon steps
can desynchronize two traces up to, notably, an arbitrary number of
(context) epsilon steps. The fact that the number of epsilon steps is
identical in both starN does not seem to be a very useful fact.

In addition, the high-level proof structure by case analysis would
appear to have errors. Consider, for example, the case where the second
star takes a smaller number of steps than the first star. This does not
rule out, as posited, that the second star cannot possibly get ahead: in
fact, with fewer steps (and fewer epsilon steps in the context) a
program could do more and  produce a strictly larger trace than the
first star.

The proof of star_improvement follows easily from two auxiliary results
that should be relatively easy, if somewhat tedious.