First steps of a state determinacy proof (program)
While the entire sequence of inversions (as in Source) is necessary to
begin to discharge some subgoals, the sequence of tactics that was used
there to take care of simple cases and contradictions will not work. The
context searches used there become very expensive in this scenario and
should be handled differently, as times do not scale gracefully.