Contradictions in program state determinism
A combination of contradictions on the pairs of opcodes generated by
each step inversion and the correspondence of state types removes a vast
majority of the many subgoals generated by the proof of state
determinism on the side of the program. Careful inspection of the order
of application of inversions and discharge tactics keeps times very
manageable.

Some nonsensical goals pertaining to the second category remain
associated to the various jumping opcodes.