Progress on program state determinism
Add a series of tactics to process the sequents that result from the
main case analysis. The idea behind all these is simple: the sequence of
inversions generates a large number of sub-goals based on instructions
in the intermediate machine. Inversion on steps of the CS semantics in
particular has a high fan-out and is cumbersome to name, but in the same
way we equated the pairs of instructions we want to equate the various
parts of the instructions in the remaining goals, where the instructions
themselves coincide. The easy way to do this in our predicament is to
have a number of small tactics look for pairs of matching assumptions in
the context (e.g., pairs of register reads, etc.) and use these,
possibly alongside other information from the context and existing
lemmas, to establish the relation between them, which is then used to
simplify the goal.

We do not wish to rely on search and pattern matching of the sequents
more than is stricly necessary at this point, and for this reason the
tactics only use pattern matching to fetch from the context the unnamed
assumptions, taking known facts as parameters. At the moment, we try to
simplify each component whenever it avails in the sequent, and make no
effort to be selective about its potential application.

A couple of simple lemmas, whose proofs are elided, are formulated.

Finally, the original micro-tactic used to unify matching operations has
been split into two: one, to discharge nonsensical equalities (its
previous use); another, to perform both equality and substitution.