Harmonize MultiSem.lockstep_simulation
Note that the proof, when the program is in the first state, follows the
same structure as that in the second state. Minor simplifications are
made to the existing proof.

On both sides, the two remaining sub-goals relate the step taken by the
unpartialized, merged state, and its relation to the partial merged
states that result.