Alternative same_step definition
The problems identified in the Composition module seem to arise from two
decoupled and potentially incompatible definitions of what it means for
a program execution to alternate turns between "program" and "context".

Ordinarily, i.e., in both Source and Intermediate PS, partial states are
defined according to whether the currently executing component is in the
program or in the context (resp. PC or CC). Turns are defined as
relative to an interface which is assumed to represent the context:
thus, is_context_component and is_program_component. The synchronization
between the two is given by partial_state, which -- always module a
context interface -- associates a CS state with a PS state of the PC or
CC form given its adherence to either is_program_component or
is_context_component, respectively, for the given context interface.

By contrast, the same_turn relation defined in the Composition module
takes only into account the syntactic constructor of the partial states,
and not the interface to which they are relative. This desynchronization
can lead to nonsensical results, given that when the reference interface
is change so do turns, from the point of view of the interface. Such
situations of confusion should be avoided.

An alternative definition of the same_turn relation folds the context
interface back in and defines turns in terms of said interface and the
usual is_program_component and is_context_component, rather than the
partial state constructors. The changes are extended to the definitions
of the "fancy stars" that incorporate turn alternation, as well as
proofs that depend on them. The changes are modest and all proofs,
insofar as they are fragmentary, can be adapted without great pains and
preserving their structure and the application of the modified lemmas:
an encouraging result, although it remains to be seen that they allow
for painless closing of the remaining gaps.