First fixes to mergeable_states
The specific conditions under which mergeable_states is designed to work
are not consistent with the definition or turns or even with the
possible combinations in which CS states can be partitioned into two,
mergeable halves. The definition is rephrased in terms of the standard
turn definitions (relative to interfaces). The usual disadvantage that
[constructor] and friends are no longer driven by syntax in typical
situations must be taken into account.

In the PS module, some results that still relied on PC vs. CC
distinctions are amended without issue.

In the Composition module, the changes are extended with a number of
more interesting situations. For starters, the problematic proof in
section ProgCtxSim should be able to go through, and most other
instances of use of the modified definitions are easily amended ([apply]
instead of [constructor], introduction of the new, added hypotheses with
the corresponding blemish in automatically generated identifiers, as
remaining in some proofs, etc. Some existing proof snippets,
specifically involving the use of mergeable_states and merged states,
are commented out, as the new definitions will require some additional
work in them if everything else works as expected.