Mergeable interfaces
Considering the mergeability of states and performing actual merges are
operations which are conceived to be meaningfully applied in very
specific conditions. To begin to express these conditions, the link with
program interfaces is established.

The new condition is added to the mergeable_states relation, new
hypotheses introduced as needed and changes propagated. Modulo the
temporary omissions in (partial) proofs introduced in the previous
commit, only the (painful) attention required by automatically named
hypotheses need to be revisited in addition to the obvious changes. No
problems are observed in the approach so far.