Add structural constraints to mergeable_states
As we have pointed out before, the standard use case of mergeable states
is to relate two projections of a CS state, each obtained through one of
the "halves" of a partition of the full program interface. Thus, one of
the halves will be a PC with respect to its half of the interface, and
the other will be a CC. Adding this restriction avoids issues associated
with the potential mergeability of registers and pc pointers.

The relation as it stands does not currently require that states be two
halves of a partition per se, but is closer to its intended usage.