mergeable_states from partial_state
Redefine mergeable_states so that a pair of partial states is mergeable
if they can be generated as partializations of a complete state that
originates from an initial state after a sequence of steps. We want to
know enough about the pairs of states that are mergeable to allow
meaningful reasoning, and this view corresponds to the more restricted
scenarios that are envisioned.

The change requires some substantial changes to existing composition
proofs; many of these are amended presently. Some auxiliary lemmas and
invariants may still be missing.