multi_match_final_states
Proof redone according to new definition. The structure varies w.r.t.
the original incarnation of the proof, including the structure of the
cases and the use of partition lemmas. It would be very interesting to
see if the old structure can be followed more closely, and what
consequences that has.