Restore partition lemma
After discussing with @arthuraa, the missing half of the both/neither
partition pair is restored. Both are currently incomplete, both have the
same fix, and one can be defined in terms of the other. Despite future
improvements, as they tend to appear in subgoals in pairs, they are
convenient to have at the moment.