Stub CtxProgSim counterpart to ProgCtxSim
Given that ProgCtxSim allows reasoning in one direction
(Program-Context) and that we may furthermore need to reason about the
opposite direction (Context-Program), a mirror module is introduced,
whose proofs must exactly parallel those of ProgCtxSim.

Discussed with Guglielmo, who also considers it sensible. Refactoring
both modules ProgCtxSim and the new CtxProgSim in their current
structure, so that both can be derived as instantiations of a common
generator, is not direct in their current form.