Dedicated section for st_starN_simulation
The missing half of the proof of st_starN_simulation relies on a
lockstep simulation different from the one that is defined in its origin
module ProgCtxSim. In fact, it is the very simulation property that
appears in its its complement CtxProgSim. The corollary
st_starN_simulation must thus follow from both simulations, not from a
single one of them.