Match initial and final states in CtxProgSim
As expected, the proofs are variations of their counterparts in
ProgCtxSim. The still unfinished proof of lockstep_simulation in the
latter should, once completed, similarly serve as the model for the
proof in CtxProgSim, which is now left empty (it makes little sense to
replicate a complex but partial proof, even if refactoring is not a
consideration).