Full case of ProgCtxSim.lockstep_simulation
The construction of the CS state and the proof of its properties is
currently verbose and makes use of a renaming tactic originally employed
in PS, but which may need to be moved elsewhere. Next, it needs to be
extended to the other cases with an eye towards finishing the entire
proof.