Remove admit on partialize in top-level proof
Add a lemma stating the equality of two partialized initial states along
the lines of a context, given that the program side stays the same. This
reasonable result depends in turn on the properties of the initial
machine state code -- work on these was put on hold when we realized
blame_program was broken.