Move blame lemma to Source PS
The blame lemma takes the place and the name of the old blame_program.
In Source, the lemma needs to take p and its well-formedness as
additional hypotheses. Moreover, the assumption on interface equality is
rephrased to make reference only to source interfaces and not to a cross
of source and interface. We pull this equality out to the top level via
preservation of interfaces in the compiler, which was previously part of
the proof of the lemma.