Abstract RSC_DC_MD (1)
Proof of concept that abstracts just the direct compiler interface;
dependent modules are simply used concretely.

Inspecting the proof we obtain a list of the definitions and theorems
from the Compiler module that RSC_DC_MD uses. Based on these, we define
a [Module Type] which copies lemma statements verbatim. We create an
instance of this module that uses exactly the definitions in S2I, but
also encapsulates them.

Next, RSC_DC_MD becomes a module parameterized by a module with the
abstract module signature for the compiler. Here we need to be explicit
about the provenance of the definitions: everything has to be phrased in
terms of the definitions and theorems of the module parameter -- not to
be confused with the definitions that are available from the context.
Some qualified renaming and tweaking of parameter passing is needed.

Finally, we create an instance of RSC_DC_MD based on the concrete (if
abstracted) Compiler module. This contains the refactored equivalent of
our previous top-level theorem.