Start abstracting Source (currently broken)
See comment: we need to pass Coq's module system the right information
so that the built-in compiler can be understood in terms of a generic
interface to the compiler and the source (and, eventually, the
intermediate language) while respecting the type signatures. I suspect
the result will resemble the usual OCaml idioms.