Abstract RSC_DC_MD (and 3): Intermediate
See comments for some technical information. Source currently presents a
dependency on Intermediate, which is probably extraneous and should
eventually be moved out, thus simplifying the nestings.

Some local changes to the top-level proof were needed beyond the usual
addition and correction of qualified module names, due to its local
reliance on low-level tactics manipulating entities of the concrete
Intermediate. For this reason the abstract interface has to be extended
with some additional exposed definitions and inversion lemmas, which
correspond to trivial proofs in the instantiation of the module
signature with the concrete instance.

The proof is now generic modulo the Common and CompCert infrastructure.