Relocate interface mergeability
In the composition assumption of the high-level proof, at least
currently, an explicit appeal to interface mergeability is required. It
does not make sense that this definition appear in one of the concrete
modules we are abstracting through those interfaces, which points to a
better placement.

Currently, only mergeable interfaces are relocated to a more suitable
location. Other related definitions remain in Intermediate.Composition,
though in many cases they may also be relocated closer to the
definitions of, e.g., memories, and so on.