First specs of global_env
For both environment and initialization code, we will want to prove
properties such as those included in this commit: we will want to know,
essentially, that they operate on a per-component basis; in other words,
the domains of the resulting maps coincide with those of the programs
they operate on. Consequently, compositional reasoning in the presence,
say, of linking, becomes natural.