Quick tweaks
Among these, remove [Admitted] keyword applied to a currently commented
out, partial lemma -- we will have to see if this needs to stay or can
be removed. This is ostensibly so that grepping gives us a quick and
accurate picture of the state of the open parts in the development.