Minor tactic refactoring
It is not as clear as in the case of state_determinism_program whether
one wants to chain together many small pieces that unify parts of the
goal. Here, there are many fewer sub-goals left, and those
corresponding, respectively, to ICall and IReturn do not have too many
points in common to the same degree that we saw on the program side.

With some easy refactoring left, this concludes the intermediate PS.