Work towards simpler initialization code
The goal is for the missing partition lemmas to have easier proofs that
do not involve painstaking analysis of relatively involved, "stateful"
(in the sense of cumulative effects) when a simpler scheme will suffice.
This seems fairly easy up to reserve_component_blocks at least. The
obvious small inefficiency being introduced by the reuse of existing
code is the need to unzip the final result.