Sketch some proofs on simplified definitions
Instead of working with the "old" definitions, we give alternative
definitions, essentially identical, with very minor differences, that
make use of the revised definitions. The properties to be proved are
almost identical, but some of the constituent results --illustrated here
by the domain preservation lemmas-- now follow directly from the new
formulation.