Matching mains
Because of the relaxed connection between a program's interface and the
presence in it of a main procedure (only at the Intermediate level; in
the Source this connection is stronger), we need to state explicitly
what it means for two programs to be "mains-compatible": that is, one if
the programs has a main iff the other does.

This condition needs to be added to the standard equality of interfaces
to complete the proof of Intermediate.interface_preserves_closedness_r:
if we replace a program inside a link with another program, the
closedness of the linking is preserved if the replacement program is
sufficiently like its predecessor. Because we are requiring more than
equality of interfaces, the name is now a bit of a misnomer.