Close top-level proof with matching mains
Given that we currently have two separate modules, Linker and Compiler,
that separately make use of matching_mains in small but related ways, a
"crossover module" S2I is added to the interface structure to represent
those components that bridge the representation of the Source and
Intermediate levels. With these, it is easy to expose the new definition
and the necessary properties to express the extended specification of
both Linker and Compiler, and using these re-finish the proof.