Enrich intermediate well-formedness of programs
As we noted while doing the proof of compositionality of the initial
states, the disconnect between the program interface and the separately
defined main procedure -- introduced around January to impose
preservation of interfaces by the compiler -- exhibits some rough edges
when we try to prove the independence of the initial states of a program
link. (Roughly, during compilation a new main component is added, which,
if exported in the interface, impedes preservation of components: this
is what motivated the change.)

The solution is to strengthen the definition of well-formedness of
programs to rule out the bogus case in the least restrictive way
possible: if a program does not define the main procedure, then it
cannot be the case that the main component is part of its interface. The
conditions under which this inclusion of the main component in an
interface are not clearly specified and may not even make sense, but we
need not be too restrictive at the moment: the weakest possible
specification that forbids the nonsensical case should suffice.

The proofs require only small additions, local to the module; everything
else remains unchanged.