GlobalEnv spec
The last proof about the initialization of global environments is given
here. The statement needs to be enriched with conditions about the
well-formedness of the intervening programs and the linkability of their
main procedures.

These changes will need to be propagated in the following series of
commits. They break the existing proofs in a good number of places; the
fixes are systematic but will be given for clarity in groups of related
changes.