Propagate specs (II): PS
Most of the work in this file takes place in the "unification" tactics
used by the proofs of state determinism: passing new parameters and
adjusting their operation according to the full specs.

The original decision to make as much context information explicit in
the form of hypothesis names can become unwieldy, and it may be the time
to consider whether to pass only the essential parameters and let
pattern matching on the goal do the rest.