Propagate specs (III): Composition, Decomposition
The main trick used here to preserve the existing proofs as much as
possible is the gradual introduction of names where none were previously
given. In particular, tactics like the inversion of predicates where new
assumptions have been introduced (notably, the constituents of PS:
initial and final state, and step) result in renumbering of implicit
assumptions, which disrupts those proofs that make use of those
uncontrolled names. The idea, then, is to locate tactics where new names
are introduced and give names only to the new assumptions. In this way,
changes to the proofs remain local. This is still a hack, but is less of
one than the direct alternative, and opens the door to sanitizing the
proofs by degrees.