Added read of a component's local buffer by another in the events
Made changes necessary to include the value read in the event.
Had to add an EqType for the values to allow comparison between events
now that a value is part of it.

This would also allow to have pointers returned or passed as argument in
the events only (it could only take integers before).

Plus, with the imports of ssrnat and eqtype, it is not even necessary to
have Pointer.eq since the equality of pointers ( (nat * nat) * Z) is
already inferred (nat and Z already have EqTypes)