Added public/private differentiation in local buffer expression
Why wouldn't components choose which part of their local buffer they can
reach ? :-)

Changed the CS and PS to adapt also.
Modified btw the definition of local to public (Values.v : block).
Breaking things is ok since we need to check that we are not putting
private informations (counters, etc) in the public local buffer.