First lemma repair attempt
As we knew, the provenance lemmas did not contain enough information in
their premises to rule out bogus executions. Here we attempt to rephrase
one of them by adding the missing information formulated in terms of
provenance from an execution starting on an initial state according to
the relevant interface. The existing proofs can be easily repaired by
adding the new premises, which are readily available in the context.

The revised lemma statement looks provable now.