Extend domain partition lemmas
The basic building block requiring use of provenance information are the
two lemmas that relate the domains of two linked interfaces (one can be
defined in terms of the other). The reparation consists in adding the
same kind of missing provenance information, relating it to existing
hypotheses and propagating it to the proofs that used them through
simple local modifications; as before, all necessary information is
available in the sequents.

Some design choices about the decomposition of states and components can
be explored in order to simplify the resulting statements and the
overhead in the revised proofs.