Base case of domm_partition
Unroll some of the overly eager case analysis (once the new information
has been added to the premises). Observe then that the extant case
analysis has no effect on the resulting cases, which are case-analyzed
duplications of one and the same proof, which itself relies only on the
well-formedness of the program and a few simple facts.