Derive well formed events and close domm_partition
The proof is finished by extracting the statements on the individual
well-formed events from the sequence of well-formedness of the events of
the corresponding traces. With this, the proof is finished with only
minor extensions to the provenance model and no real effect on existing
proofs depending on the result, as was expected.