Proof of mergeable_stacks_partition
As the remaining partition results foreseeably do, this proof proceeds
by an administratively slightly involved induction of the same sort as
that used by the more elementary result domm_partition. Beyond the usual
matters of proof style, some opportunities for refactoring present
themselves, and some will become more apparent as the patterns repeat in
the remaining proofs of the same family of results.