Fix typo/bug in mergeable_stack_frames
In the _second constructor, the first component was erroneously used on
both sides of the relation. Because this was forcing a sort of bogus but
"mostly correct" unification of the two components while leaving the
other unspecified, the error had no effect in existing proofs. This
raises the point of why separating the two components which indeed have
to coincide and leave this coincidence to a separate premise, when
unification would do the job out of the box.