Lemma selection in ProgCtxSim
One significant aspect of making the proof of lockstep_simulation
modular (beyond folding in case differences under a common framework) is
selecting the right constructor of CS.step based on the instruction that
is executing at a certain point in each of the cases of the main, case
analyzing section of the proof. For this we cannot rely on the goal of
the proof alone, and instead use a bit of Ltac to inspect the context
and select the appropriate constructor to apply.