Another ProgCtxSim case, another tentative proof
As we note in the comments, the simple proof script that works in the
first few cases with very small changes, used without further
adjustments, leads to some sub-goals that are not immediately solvable
from the context as in previous cases.

A possible alternative explored here purports to use the information in
the context, notably provenance and stepping hypotheses, to move

 - from a CS.step on the program, p, and a context, p', that is
   compatible with and replaces c,

 - to a new CS.step on the program, p, and the original context, c,
   built based on the state before the step and on the step proper,
   which is on p's domain.

From this, the proof for the ILoad case is simplified in various ways
and avoids sub-goals where the solution is not straightforward.