Definition for memory partialization
In the same way that stacks are partialized by applying a dedicated
partialization function that hides the details of its implementation, we
add the missing function that performs the analogous process for
memories. This makes code more readable and will be of use to write
lemmas on partialized memories. With the proofs as they currently stand,
the introduction of the change involves some unfolding of the definition
at appropriate points, but no further changes are needed.

A further abstraction may be possible by having these partializer
functions operate on a piece of the state (stack, memory) and, instead
of the domain of a program interface, on the program interface itself.