IStore case in ProgCtxSim.lockstep_simulation
Now we can tackle one of the more interesting cases, where the operation
being executed has a visible effect on one of the more central
components of the stage, in particular memory. Here we need to be a bit
more careful to get the right information into the context, in
particular by piecing together states using two separate CS programs and
their executions, both always involving the program p, as well as parts
of the two context in scope, the original c and its contextual
doppelganger p'. The parts are accessed by using the composition lemmas
stated in previous commits, and beyond some reordering of the lemmas
that have to be rewritten a bit later than in previous, trivial cases
(which did not require so much thought up front, imposing fewer
restrictions on the necessities of the intermediate results), as the
appropriate information is revealed from the definitions, the proof
strategy proceeds without significant changes.