Stack operations in ICall case of ProgCtxSim
By applying the posited rewrite lemmas we complete the ICall case almost
to complete satisfaction, modulo some refactoring and automation of
extraction of parts of the proof (from goals, hypotheses) to serve as
parts of the proof script. It remains to ensure the proper form of the
stack lemmas, as well as the proper derivation of the provenance
information on which they depend; none of these seem particularly
challenging.