Mostly complete IReturn case of ProgCtxSim
Similar comments as those applicable to the ICall case. Composition of
stepping and provenance results, as well as some analysis of local data
structures, are the (relatively small) missing pieces at this point.
Once these are provided the lemma and section should be finished, and
the results easily extended to the complementary CtxProgSim.