Tentative difficulty rating for CtxProgSim
This should be closely modeled after the counterpart in ProgCtxSim. The
relation between both lockstep_simulation theorems is probably similar
to that between the pair of theorems that state the determinacy of CS
steps in the program and in the context. It may be worthwhile for the
intrepid prover to check the state of the missing bits in ProgCtxSim --
I expect to close those shortly.