Update proof of MultiSem.lockstep_simulation
Sections of the proof that were commented out are now restored with the
necessary changes; some new and changed goals are resolved and
contradictory cases discharged. In addition, inverted hypotheses are
cleared to try to keep the contexts relatively light. The more
interesting missing case on the first sub-case of the main pair of
symmetrical cases should proceed by a fairly standard, if verbose, case
analysis that is left to a future commit.