Permute order of goals in MultiSem simulation
Because the two interesting goals are symmetrical, it should not affect
the proof to do case analysis up front and discard the bad cases at the
top. Moreover, this will allow sharing information between the two
conjuncts of the goal at each sub-goal, which has a chance of making a
potential proof strategy easier to apply successfully (we shall find out
soon).