Another single-memory partialization lemma
Unlike the existing lemmas that relate two partially equal memories, the
new lemmas express the partial equalities after an operation on a single
memory. The naming of the groups of lemmas differs, but may be similar
enough that making the groups more clearly identifiable from names is a
good idea. Easy proofs likely follow from the existing lemmas in their
strong versions.