Proof sketch of turn change impossibility lemma
Outline the proof of the first of the series of lemmas that describe
situations in which turn changes are impossible. The remaining lemmas
should follow very similar patterns and are not too complicated. To get
to the final contradiction, we need to make use of provenance
information on the runs as the only salient detail.