Mark problematic lemmas
The statements of the two lemmas relating two executions with the same
traces and number of steps are too weak and will not hold as they stand.
These are used in the proof of [threeway_multisem_mt_starN_simulation],
whose statement does seem reasonable but whose proof will need to be
revised.