State extension of state determinism to st_starN
Observe that a partial step is always deterministic, whether in the
program or in the context, and a sequence of steps remains deterministic
(from the point of view of a partial program) until a change of turn
occurs. From the point of view of the program, this is direct. From the
point of view of the context, the shared trace is the only necessary
equalizer, which gives us the current component -- the context-local
memory and stack do not change.