Revised proof strategy for threeway multisem
The inductive case of threeway_multisem_st_starN_simulation is rephrased
for clearer reasoning and encapsulation of pesky case analysis (the base
case was already established and there is no need to change this now). I
do not think I am missing anything in this new formulation, but of
course the proof needs to be finished.