Comments on threeway simulation, couple of tweaks
I think the theorem is provable by the two fairly simple admits that are
left in the new strategy, unless some small detail is overlooked. It
does not appear to me this should be difficult to finish, so I have
assigned it a rating and a small warning (this is probably closer to a
grade 2, but there might be some annoying detail hidden there, although
I do not see any now). I also suspect a similar strategy may reap
benefits especially throughout PartialComposition.