Stub multi-turn simulation
Just as we are able to simulate a single turn from one run into another,
parallel run, it should be possible to extend this to multiple turns.
This turns out to be simple if we admit the extension to the turn
change, and could in turn help us in other proofs later in the
development.