New strategy for threeway multisem
The inductive case of threeway_multisem_mt_starN_simulation admits an
easy formulation by using the simpler simulations on the various
components of a multi-turn star. These make no use of those admitted
statements that we identified as incorrect and require only a proof
that, in our particular scenario, the execution of a series of turns
is determinate.