Comment out unused helpers
One of the potential advantages of the new proof strategy used in the
threeway simulations is that it relies on fewer and simpler lemmas, and
none of those statements whose formulations were observed to be flawed.

It would be interesting to see if we can get rid of the remaining two
helpers, since those require the addition of provenance information that
will need to extend to the threeway simulations if they stay in place.