Alt proof idea for initial_states_mergeability
The partial proof of initial_states_mergeability is replaced here by
another, cleaner proposal based on compositionality of initial machine
states.

An important point is that there are still some small but important
missing pieces in this proof, and these missing pieces are closely
related to those in a related lemma, merged_initial_states.

In both cases, a first PS initial state is related to a second CS
initial state which partializes to the first. If the PS state is on a
partial program p, inverting the PS.partial_state relation yields a
piece of CS provenance information on the linking of p with a
complementary p' satisfying the context interface.

The impediment here seems to be that, at the intermediate level, the
main procedure is not part of the interface, and so there is a
disconnect between the presence (or not) of a main procedure in the
context, which would otherwise be propagated by PS.partial_state and
reflected in the provenance information.

In short, the main procedure hack comes back to haunt us yet again. It
is likely that at least a partial fix is necessary for a satisfactory
solution to these pieces of the development.